requirements.txt
Coffin==0.3.8.dev
Django==1.5.1
Fabric==1.6.0
Jinja2==2.6
Markdown==2.3.1
Pillow==2.0.0
PyYAML==3.10
South==0.7.6
Twisted==13.0.0
amqp==1.0.10
anyjson==0.3.3
beautifulsoup4==4.1.3
billiard==2.7.3.23
## FIXME: could not find svn URL in dependency_links for this package:
bulbs==0.3.13-20130320
celery==3.0.17
coverage==3.6
csat==0.1a
cssmin==0.1.4
cssutils==0.9.10b1
distribute==0.6.35
django-assets==0.8
django-celery==3.0.17
django-crispy-forms==1.2.3
django-debug-toolbar==0.9.4
django-polymorphic==0.4.0
httplib2==0.8
ipython==0.13.1
jsmin==2.0.2-1
kombu==2.5.8
omnijson==0.1.2
paramiko==1.10.0
psycopg2==2.4.6
pycrypto==2.6
python-dateutil==1.5
pytz==2013b
six==1.3.0
txWS==0.7.1
webassets==0.8
wsgiref==0.1.2
zope.interface==4.0.5
